Refactor PeerAuthentication.
